[add] added success message to Login, Register and Post-product form pages, also disabled the use of context api, commented the whole page out and now using use state for my state management
