+++ app/auth/register/page.jsx
@@ -3,12 +3,18 @@ import styles from "../Auth.module.css"
 import { useForm } from "react-hook-form"
 import * as yup from "yup"
 import { yupResolver } from "@hookform/resolvers/yup"
-import { useAppContext } from "@/app/context/ContextProvider"
 import { useRouter } from "next/navigation"
 import { authentication } from "@/app/utils/firebaseConfiguration"
 import { createUserWithEmailAndPassword } from "firebase/auth"
+import { useState } from "react"
 
 const Register = () => {
+
+    const [isLoading, setIsLoading] = useState(false)
+    const [err, setErr] = useState(false)
+    const [errMsg, setErrMsg] = useState()
+    const [successMsg, setSuccessMsg] = useState("")
+    const [success, setSuccess] = useState(false)
 
     const schema = yup.object().shape({
         username: yup.string().required("Username is required!"),
@@ -22,19 +28,20 @@ const Register = () => {
     })
 
     const {push} = useRouter()
-    const { isLoading, setIsLoading, 
-         handleErrMsg, err, setErr, errMsg, } = useAppContext()
-
-        // console.log("Authentication "+authentication, "AuthCurrentUser "+authentication.currentUser, "stateUser "+currentUser)
 
     const handleRegister = async (data)=>{
         try {
             setIsLoading(true)
             setErr(false)
-            // const res = await creatAccount(data.email, data.password)
+            setSuccess(false)
             const res = await createUserWithEmailAndPassword(authentication, data.email, data.password)
-            res?push("/post-product"):""
-            setIsLoading(false)
+            if (res) {
+                setIsLoading(false)
+                setSuccess(true)
+                setSuccessMsg("Registration-successful")
+                push("/post-product")
+            }
+            
         }catch (error) {
             setIsLoading(false)
             setErr(true)
@@ -44,11 +51,54 @@ const Register = () => {
         // setIsLoading(false)
     }
 
+    const handleErrMsg= (err) =>{
+        switch (err.code) {
+            case "auth/email-already-exists":
+                setErrMsg("Email-already-exists")
+            break;
+            case "auth/internal-error":
+                setErrMsg("Internal-error")
+            break;
+            case "auth/invalid-email":
+                setErrMsg("Invalid-email")
+            break;
+            case "auth/invalid-password":
+                setErrMsg("Invalid-credentials")
+            break;
+            case "auth/user-not-found":
+                setErrMsg("User-not-found")
+            break;
+            case "auth/network-request-failed":
+                setErrMsg("Network-request-failed")
+            break;
+            case "auth/wrong-password":
+                setErrMsg("Wrong-credentials")
+            break;
+            case "auth/too-many-requests":
+                setErrMsg("Too-many-requests")
+            break;
+            case "auth/email-already-in-use":
+                setErrMsg("Email-already-in-use")
+            break;
+            case "connectStorageEmulator is not defined":
+                setErrMsg("Network-error")
+            break;
+            case "email is not defined":
+                setErrMsg("Wrong-credentials")
+            break;
+            
+            default:
+                setErrMsg("Something-went-wrong")
+            break;
+        }
+    }
+
     return (<>
         <div className={styles.form_wrapper}>
             <form className={styles.form} onSubmit={handleSubmit(handleRegister)}>
                 <h1>Register</h1>
                 {err&&<p className="errMsg">{errMsg}</p>}
+                {success&&<p className="successMsg">{successMsg}</p>}
                 <div className={styles.inputCategory}>
                 <label htmlFor="">
                         <h4>Username</h4>
